requirements.txt
flask
pyTelegramBotAPI
google-generativeai>=0.8.3
gunicorn
